requirements.txt
Pillow
PySide6
qrcode
